requirements.txt
pyaudio
pyacoustid
requests
pygame
python-mpd2